0.3.2
Style updates including size increases to make things easier to see and to take on the theme of SwarmUI in all modals. Some are hard to see in the text boxes, changes coming for that.

Moved around a few items in settings modal for future releases.

Updated some default model settings for better performance.